Added socials to player pages. Misc UI work site wide
