+++ src/App.css
@@ -230,7 +230,7 @@ body {
 .clip-title {
     color: white;
     font-family: "Scout Condensed Bold";
-    font-size: 2em;
+    font-size: 3em;
     margin: auto;
     text-align: center;
 }
