+++ src/pages/Community.js
@@ -29,9 +29,9 @@ const Community = () => {
             {community.length > 0 &&
             <>
             <div className='grid grid-cols-12 gap-4 m-4 2xl:grid-cols-5'>
-                {community.map( player => {
+                {community.map( (player, i) => {
                         return (
-                            <PlayerCard player={player} />
+                            <PlayerCard key={i} player={player} />
                         )
                 })}
             </div>
